# attendance-app-webapi/docs/swagger.yaml
basePath: /api
definitions:
...
  LeaveValidationRequest:
    properties:
      approverNotes:
        description: Optional notes from the approver
        example: Approved based on medical documentation
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.LeaveRequestStatus'
        description: Status to set for the leave request (APPROVED, REJECTED)
        example: APPROVED
    required:
    - status
    type: object
...
  models.LeaveRequestStatus:
    enum:
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-varnames:
    - LeavePending
    - LeaveApproved
    - LeaveRejected
  models.LeaveRequestSwagger:
    properties:
      ApproverID:
        type: integer
      ApproverNotes:
        type: string
      AttachmentURL:
        type: string
      CreatedAt:
        type: string
      EndDate:
        type: string
      ID:
        type: integer
      LeaveType:
        $ref: '#/definitions/models.LeaveType'
      Reason:
        type: string
      StartDate:
        type: string
      Status:
        $ref: '#/definitions/models.LeaveRequestStatus'
      UpdatedAt:
        type: string
      UserID:
        type: integer
    type: object
  models.LeaveType:
    enum:
    - SICK
    - PERMIT
    type: string
    x-enum-varnames:
    - Sick
    - Permit
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API service for digital attendance system with role-based access control.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Digital Attendance API
  version: "1.0"
paths:
...
  /user/leave:
    post:
      consumes:
      - multipart/form-data
      description: Submit a new leave request with supporting documentation
      parameters:
      - description: Supporting document (JPG, JPEG, PNG, PDF, max 5MB)
        in: formData
        name: attachment
        required: true
        type: file
      - description: Type of leave (SICK, PERMIT)
        in: formData
        name: leaveType
        required: true
        type: string
      - description: Start date of leave (YYYY-MM-DD)
        in: formData
        name: startDate
        required: true
        type: string
      - description: End date of leave (YYYY-MM-DD)
        in: formData
        name: endDate
        required: true
        type: string
      - description: Reason for leave request
        in: formData
        name: reason
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LeaveRequestSwagger'
        "400":
          description: Invalid request payload, dates, or attachment
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit leave request
      tags:
      - leave
  /user/leave/my-requests:
    get:
      consumes:
      - application/json
      description: Get all leave requests submitted by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LeaveRequestSwagger'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my leave requests
      tags:
      - leave
  /user/leave/subordinates:
    get:
      consumes:
      - application/json
      description: Get all leave requests submitted by users who report to the current
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LeaveRequestSwagger'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get subordinate leave requests
      tags:
      - leave
  /user/leave/validate/{id}:
    put:
      consumes:
      - application/json
      description: Approve or reject a leave request as a supervisor
      parameters:
      - description: Leave request ID
        in: path
        name: id
        required: true
        type: string
      - description: Leave validation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LeaveValidationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LeaveRequestSwagger'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Not the supervisor
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Leave request or user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Validate leave request
      tags:
      - leave
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: For testing, you can use this admin test token (valid for 1 year):<br>Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOiJhZG1pbiIsImV4cCI6MTc5MzMyODc5OX0.sKEg9JeUxDqHyD4vGwAb6pk9iLrebJfBpTJGPRnMSrY
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"